Flow Completed / Update
Flow complete and now matches current flow requested by Barry.
Poly Line activity and Maker Activity allow a user to mark an image they
have just taken.

Current Bug list:
When run on a Tablet, the  orientation is wrong for the camera code so
code needs to be written to take the fact in android that tablets X,Y,Z
are different than a phones.

Data is not saved for Poly line or Marked image  :

In general the app internal uses the string describing actions as the
control points for the flow of the app, this should be changed to states
with int codes for each state instead.

Unknown bug on Andorid Ace phones were the program crashes / or turns
black when the user tries to view the image

Improvements needed:

MAP component needs to display all observations made.
MAP ideally will connect with the backend and show other peoples
observations.
GPS code needs to be improved
Angle of QualityCamera activity needs to be recorded and the new camera
should only allow a picture to be taken when the smartphone is facing
the horizon down to 15 degrees  or so.